backend/pyproject.toml
[project]
name = "digital-fte"
version = "0.1.0"
description = "Digital Full-Time Employee - AI-powered business automation with local-first privacy"
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    # Claude Code CLI is used directly (installed separately via npm)
    # No Python SDK needed - uses subprocess to call 'claude' command
    "watchdog>=4.0.0",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.116.0",
    "playwright>=1.41.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "croniter>=2.0.1",
    "tenacity>=8.2.3",
    "tweepy>=4.14.0",
    "fpdf2>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
digital-fte = "digital_fte.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
